# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/classroom": {
            "get": {
                "tags": [
                    "classroom"
                ],
                "summary": "Get all classroom",
                "description": "Get all classroom under the logged in admin's branch",
                "operationId": "bb75c8e478344960925aff714634f5ef",
                "responses": {
                    "200": {
                        "description": "List of all classroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "branch_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "01022001"
                                                    },
                                                    "max_student": {
                                                        "type": "integer",
                                                        "example": 30
                                                    },
                                                    "student_quantity": {
                                                        "type": "integer",
                                                        "example": 21
                                                    },
                                                    "active_status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-05-05 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-05-05 12:00:00"
                                                    },
                                                    "branch": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "main-branch"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "classroom"
                ],
                "summary": "Create new classroom",
                "description": "Create new classroom and return created data (classroom name format:[first two digits-branch no,\n            next two digits-floor no, next two digits-room no]->regex:/^(10|0[1-9])\\d{2}(0[1-9]|[1-9][0-9])$/)",
                "operationId": "createClassroom",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "name",
                                    "max_student"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "010220"
                                    },
                                    "max_student": {
                                        "type": "integer",
                                        "example": 35
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "branch_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "010220"
                                                },
                                                "max_student": {
                                                    "type": "integer",
                                                    "example": 35
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2021-05-05 12:00:00"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2021-05-05 12:00:00"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/classroom/{id}": {
            "put": {
                "tags": [
                    "classroom"
                ],
                "summary": "Update a classroom",
                "description": "Update a classroom by id",
                "operationId": "updateclassroom",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "classroom object that needs to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_student",
                                    "active_status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "010220"
                                    },
                                    "max_student": {
                                        "type": "integer",
                                        "example": 35
                                    },
                                    "active_status": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "classroom updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "branch_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "010220"
                                                },
                                                "max_student": {
                                                    "type": "integer",
                                                    "example": 35
                                                },
                                                "active_status": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "classroom"
                ],
                "summary": "Delete a classroom",
                "description": "Delete a classroom by its ID.",
                "operationId": "deleteClassroom",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the classroom",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}